+++ GeneralsMD/Code/GameEngine/Include/GameLogic/Module/SupplyWarehouseDockUpdate.h
@@ -68,6 +68,7 @@ public:
 	virtual Bool action( Object* docker, Object *drone = NULL );	///<For me, this means identifying who is docking and either taking Boxes away or giving them
 
 	Int getBoxesStored() const { return m_boxesStored; }
+	virtual Int getSupplyCashValue( void ) const;
 
 	void setCashValue( Int cashValue );